Make clang -Wextra happy

+++ src/beatmap.h
@@ -42,21 +42,21 @@ struct action {
  * file name, and stores the absolute path to it in *map.
  * Returns the length of the path stored, or zero on failure.
  */
-int find_beatmap(char *base, char *partial, char **map);
+size_t find_beatmap(char *base, char *partial, char **map);
 
 /**
  * Parse a beatmap file (*.osu) into an array of hitpoint structs pointed to by 
  * **points and a metadata struct.
  * Returns the number of points parsed and stored.
  */
-int parse_beatmap(char *file, struct hitpoint **points,
+size_t parse_beatmap(char *file, struct hitpoint **points,
 	struct beatmap_meta **meta);
 
 /**
  * Parses a total of count hitmapts from **points into **actions.
  * Returns the number of actions parsed and stored, which should be count * 2.
  */
-int parse_hitpoints(int count, int columns, struct hitpoint **points,
+int parse_hitpoints(size_t count, size_t columns, struct hitpoint **points,
 	struct action **actions);
 
 /**
